% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_namespaces.R
\name{list_namespaces}
\alias{list_namespaces}
\title{List available namespaces (data sources)}
\usage{
list_namespaces()
}
\value{
A character vector of namespace identifiers.
}
\description{
List available namespaces (data sources)
}
\examples{
\dontrun{
library(bedrockbio)
list_namespaces()
}

}
