% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tables.R
\name{list_tables}
\alias{list_tables}
\title{List available tables, optionally filtered to one namespace}
\usage{
list_tables(namespace = NULL)
}
\arguments{
\item{namespace}{If given, return only that namespace's tables; otherwise
all tables.}
}
\value{
A character vector of fully-qualified table identifiers.
}
\description{
List available tables, optionally filtered to one namespace
}
\examples{
\dontrun{
library(bedrockbio)
list_tables("ukb_ppp")
}

}
